# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MoneyPrinterTurbo",
    "description": "<a href='https://github.com/harry0703/MoneyPrinterTurbo'>https://github.com/harry0703/MoneyPrinterTurbo</a>",
    "version": "1.3.7"
  },
  "paths": {
...
    "/api/v1/tasks": {
      "get": {
        "tags": [
          "V1"
        ],
        "summary": "Get all tasks",
        "operationId": "get_all_tasks_api_v1_tasks_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 10,
              "title": "Page Size"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/episodes/{episode_id}": {
      "get": {
        "tags": [
          "V1"
        ],
        "summary": "Get one episode",
        "operationId": "get_episode_api_v1_episodes__episode_id__get",
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Episode / task ID",
              "title": "Episode Id"
            },
            "description": "Episode / task ID"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "V1"
        ],
        "summary": "Move an episode to another project",
        "operationId": "update_episode_api_v1_episodes__episode_id__patch",
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Episode / task ID",
              "title": "Episode Id"
            },
            "description": "Episode / task ID"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EpisodeUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EpisodeDetailData": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "topic": {
            "type": "string",
            "title": "Topic",
            "default": ""
          },
          "script": {
            "type": "string",
            "title": "Script",
            "default": ""
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params",
            "default": {}
          },
          "run_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Run Data",
            "default": {}
          },
          "state": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "progress": {
            "type": "integer",
            "title": "Progress",
            "default": 0
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "project_id"
        ],
        "title": "EpisodeDetailData"
      },
      "EpisodeDetailResponse": {
        "properties": {
          "status": {
            "type": "integer",
            "title": "Status",
            "default": 200
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "default": "success"
          },
          "data": {
            "$ref": "#/components/schemas/EpisodeDetailData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "EpisodeDetailResponse"
      },
...
      "EpisodeUpdateRequest": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          }
        },
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "EpisodeUpdateRequest"
      },
      "EpisodeUpdateResponse": {
        "properties": {
          "status": {
            "type": "integer",
            "title": "Status",
            "default": 200
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "default": "success"
          },
          "data": {
            "type": "null",
            "title": "Data"
          }
        },
        "type": "object",
        "title": "EpisodeUpdateResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TaskListData": {
        "properties": {
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskStatusData"
            },
            "type": "array",
            "title": "Tasks"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "tasks",
          "total",
          "page",
          "page_size"
        ],
        "title": "TaskListData",
        "description": "分页任务列表结构。"
      },
      "TaskListResponse": {
        "properties": {
          "status": {
            "type": "integer",
            "title": "Status",
            "default": 200
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "default": "success"
          },
          "data": {
            "$ref": "#/components/schemas/TaskListData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "TaskListResponse",
        "description": "任务列表使用独立响应模型，避免与单任务查询混用文档结构。",
        "example": {
          "data": {
            "page": 1,
            "page_size": 10,
            "tasks": [
              {
                "progress": 50,
                "state": 4,
                "task_id": "6c85c8cc-a77a-42b9-bc30-947815aa0558"
              }
            ],
            "total": 1
          },
          "message": "success",
          "status": 200
        }
      },
...
      "TaskStatusData": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "state": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "progress": {
            "type": "integer",
            "title": "Progress",
            "default": 0
          },
          "videos": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Videos"
          },
          "combined_videos": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Combined Videos"
          },
          "failed_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failed Stage"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "cross_post_state": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "pending",
                  "processing",
                  "complete",
                  "failed"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Cross Post State"
          },
          "cross_post_results": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cross Post Results"
          },
          "cross_post_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cross Post Error"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "task_id"
        ],
        "title": "TaskStatusData",
        "description": "任务查询对外保证的稳定字段；历史和扩展字段继续原样透传。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}